+++ pkg/mimir/modules_test.go
@@ -8,9 +8,11 @@ package mimir
 import (
 	"net/http/httptest"
 	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/gorilla/mux"
+	"github.com/grafana/dskit/flagext"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/weaveworks/common/server"
@@ -155,6 +157,58 @@ func TestMimir_InitRulerStorage(t *testing.T) {
 			} else {
 				assert.Nil(t, mimir.RulerStorage)
 			}
+		})
+	}
+}
+
+func TestMultiKVSetup(t *testing.T) {
+	dir := t.TempDir()
+
+	for target, checkFn := range map[string]func(t *testing.T, c Config){
+		All: func(t *testing.T, c Config) {
+			require.NotNil(t, c.Distributor.DistributorRing.KVStore.Multi.ConfigProvider)
+			require.NotNil(t, c.Ingester.LifecyclerConfig.RingConfig.KVStore.Multi)
+			require.NotNil(t, c.StoreGateway.ShardingRing.KVStore.Multi.ConfigProvider)
+			require.NotNil(t, c.Compactor.ShardingRing.KVStore.Multi.ConfigProvider)
+		},
+
+		Ruler: func(t *testing.T, c Config) {
+			require.NotNil(t, c.Ruler.Ring.KVStore.Multi.ConfigProvider)
+		},
+
+		AlertManager: func(t *testing.T, c Config) {
+			require.NotNil(t, c.Alertmanager.ShardingRing.KVStore.Multi.ConfigProvider)
+		},
+
+		Distributor: func(t *testing.T, c Config) {
+			require.NotNil(t, c.Ingester.LifecyclerConfig.RingConfig.KVStore.Multi.ConfigProvider)
+		},
+
+		Ingester: func(t *testing.T, c Config) {
+			require.NotNil(t, c.Ingester.LifecyclerConfig.RingConfig.KVStore.Multi.ConfigProvider)
+		},
+	} {
+		t.Run(target, func(t *testing.T) {
+			prepareGlobalMetricsRegistry(t)
+
+			cfg := Config{}
+			flagext.DefaultValues(&cfg)
+			// Set to 0 to find any free port.
+			cfg.Server.HTTPListenPort = 0
+			cfg.Server.GRPCListenPort = 0
+			cfg.Target = []string{target}
+
+			// Must be set, otherwise MultiKV config provider will not be set.
+			cfg.RuntimeConfig.LoadPath = filepath.Join(dir, "config.yaml")
+
+			c, err := New(cfg)
+			require.NoError(t, err)
+
+			_, err = c.ModuleManager.InitModuleServices(cfg.Target...)
+			require.NoError(t, err)
+			defer c.Server.Stop()
+
+			checkFn(t, c.Cfg)
 		})
 	}
 }