+++ CHANGELOG.md
@@ -421,6 +421,7 @@
 * [BUGFIX] Query-frontend: worker's cancellation channels are now buffered to ensure that all request cancellations are properly handled. #741
 * [BUGFIX] Compactor: compactor should now be able to correctly mark blocks for deletion and no-compaction, if such marking was previously interrupted. #1015
 * [BUGFIX] Overrides-exporter: successfully startup even if runtime config is not set. #1056
+* [BUGFIX] Multi-KV: runtime config changes are now propagated to all rings, not just ingester ring. #1047
 
 ### Mixin (changes since `grafana/cortex-jsonnet` `1.9.0`)